backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
pdf2image>=1.16.0
opencv-python>=4.8.0
numpy>=1.24.0
pydantic>=2.0.0
ultralytics>=8.0.0
